requirements.txt
discord
python-dotenv
pymongo
dnspython
requests
python-dateutil
